refactor: add product response dto
